fix logging in rsync-backup

+++ useful/deployed/backup-nfs/rsync-backup.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 # thanks to https://linuxconfig.org/how-to-create-incremental-backups-using-rsync-on-linux
 # A script to perform incremental backups using rsync
@@ -15,7 +15,7 @@ readonly LATEST_LINK="${BACKUP_DIR}/latest"
 BACKUP_PATH="${BACKUP_DIR}/${DATETIME}"
 
 echo SOURCE PATH is $SOURCE_DIR
-echo BACKUP PATH is $BACKUP_PATH
+echo BACKUP DIR is $BACKUP_DIR
 
 # check for a previous incompleted run by looking for the file called
 # ${BACKUP_DIR}/${DATETIME}.in-progress
@@ -35,13 +35,15 @@ mkdir -p "${BACKUP_PATH}"
 # I had to stop using -a which would give a perfect archive.
 # The thousands of symlinks in my photo albums seemed to cause rsync lockups. 
 # Options -rptgov give the same as -a except ignore symlinks
-rsync -rptgov --delete \
+rsync -rptgo --delete \
   "${SOURCE_DIR}/" \
+  --info=SKIP3,STATS,FLIST,NAME \
   --link-dest "${LATEST_LINK}" \
   --exclude=".cache" \
-  "${BACKUP_PATH}"
+  "${BACKUP_PATH}" \
+  | grep -v "skipping non-regular"
 
-unlink "${LATEST_LINK}"
+if test -d ${LATEST_LINK} ; then unlink "${LATEST_LINK}"; fi
 ln -s "${BACKUP_PATH}" "${LATEST_LINK}"
 
 # Verify completion of this backup
